Refactored project setup
