+++ src/rust/ide/src/controller.rs
@@ -20,12 +20,15 @@ pub mod graph;
 pub mod ide;
 pub mod module;
 pub mod text;
+pub mod project;
 pub mod visualization;
 pub mod searcher;
 
 pub use graph::Handle           as Graph;
 pub use graph::executed::Handle as ExecutedGraph;
+pub use ide::Handle             as Ide;
 pub use module::Handle          as Module;
+pub use project::Handle         as Project;
 pub use text::Handle            as Text;
 pub use visualization::Handle   as Visualization;
 pub use searcher::Searcher;
