+++ src/rust/ide/view/src/ide.rs
@@ -1,0 +1,28 @@
+//
+// use crate::prelude::*;
+//
+// use crate::project;
+//
+//
+//
+// ensogl::define_endpoints! {
+//     Output {
+//         create_new_project_requested ()
+//     }
+// }
+//
+//
+// #[derive(Clone,CloneRef,Debug)]
+// struct Model {
+//     project : project::View,
+// }
+//
+// #[derive(Clone,CloneRef,Debug)]
+// pub struct View {
+//     frp   : Frp,
+//     model : Model,
+// }
+//
+// impl View {
+//
+// }
